revamped control: new classification algorithm, new system state update, changed everything to float

+++ src/lsm6dsv.hpp
@@ -12,7 +12,7 @@
 struct accelerations
 {
   /// \brief 3 axis acceleration data
-  double x, y, z;
+  float x, y, z;
 };
 
 /// \brief Class abstraction of functions required to use LSM6DSV_IMU
